add request message hashing before signing
